minor: extracts InvalidAutoAddTargetException from TypesPreprocessor
